Moved toolbar management in common fragment
