+++ app/src/main/java/it/danieleverducci/nextcloudmaps/fragments/GeofavoritesFragment.java
@@ -1,9 +1,13 @@
 package it.danieleverducci.nextcloudmaps.fragments;
 
+import static android.view.View.GONE;
+import static android.view.View.VISIBLE;
+
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
@@ -13,10 +17,13 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.widget.AppCompatImageButton;
 import androidx.appcompat.widget.AppCompatImageView;
+import androidx.appcompat.widget.SearchView;
+import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.google.android.material.card.MaterialCardView;
 import com.nextcloud.android.sso.exceptions.NextcloudFilesAppAccountNotFoundException;
 import com.nextcloud.android.sso.exceptions.NoCurrentAccountSelectedException;
 import com.nextcloud.android.sso.helper.SingleAccountHelper;
@@ -40,6 +47,9 @@ public abstract class GeofavoritesFragment extends Fragment {
     private final String TAG = "GeofavoritesFragment";
 
     protected GeofavoritesFragmentViewModel mGeofavoritesFragmentViewModel;
+    private View toolbar;
+    private View homeToolbar;
+    private SearchView searchView;
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -70,6 +80,33 @@ public abstract class GeofavoritesFragment extends Fragment {
         View userBadgeContainer = view.findViewById(R.id.user_badge_container);
         userBadgeContainer.setOnClickListener(v -> showSwitchAccountDialog());
 
+        // Setup toolbar/searchbar
+        toolbar = view.findViewById(R.id.toolbar);
+        homeToolbar = view.findViewById(R.id.home_toolbar);
+
+        searchView = view.findViewById(R.id.search_view);
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String query) {
+                onSearch(query);
+                return false;
+            }
+        });
+
+        searchView.setOnCloseListener(() -> {
+            if (toolbar.getVisibility() == VISIBLE && TextUtils.isEmpty(searchView.getQuery())) {
+                updateToolbars(true);
+                return true;
+            }
+            return false;
+        });
+
+        homeToolbar.setOnClickListener(v -> updateToolbars(false));
         // Set user badge (async)
         Handler h = new Handler();
         h.post(() -> {
@@ -152,6 +189,15 @@ public abstract class GeofavoritesFragment extends Fragment {
                 .setNegativeButton(android.R.string.no, (dialog, id) -> dialog.dismiss());
         AlertDialog ad = builder.create();
         ad.show();
+    }
+
+    private void updateToolbars(boolean disableSearch) {
+        homeToolbar.setVisibility(disableSearch ? VISIBLE : GONE);
+        toolbar.setVisibility(disableSearch ? GONE : VISIBLE);
+        if (disableSearch) {
+            searchView.setQuery(null, true);
+        }
+        searchView.setIconified(disableSearch);
     }
 
 }
