+++ app/src/main/res/layout/app_toolbar.xml
@@ -19,7 +19,10 @@
         <androidx.appcompat.widget.SearchView
             android:id="@+id/search_view"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:layout_marginEnd="30dp"
+            android:background="@color/translucent">
         </androidx.appcompat.widget.SearchView>
 
     </androidx.appcompat.widget.Toolbar>
